backend_service/requirements.txt
Flask==2.1.4
opentelemetry-sdk==1.5.0
opentelemetry-exporter-jaeger-thrift==1.5.0
opentelemetry-instrumentation-flask==1.5.0
opentelemetry-instrumentation-requests==1.5.0
opentelemetry-sdk-resources==1.5.0
opentelemetry-sdk-metrics==1.5.0
prometheus-flask-exporter==0.18.1
requests==2.27.1